configure: support syncing from a remote, using remote stage
this uses a tempdir as a root for syncing repos because the builder
containers do not run as root and can't just use the host repos (also
this wouldn't work well on cross). can't use masterdirs because the
cross setup is weird there too.

also: be explicit about host arch even if native